Porting to Solaris... This compiles, I have yet to do any real testing. Once this is checked in, I will make sure these changes don't break Linux/FreeBSD.
configure.ac:
    Adding socket,nsl,rt libraries
    Adding check for paths.h header file

ErrLogSyslog.c:
    Define LOG_AUTHPRIV and LOG_FTP if they are not defined.

Options.c:
conf.c:
    Adding cast from char to int for all isspace/isdigit args.

readpassphrase.c:
    remove cdefs.h.
    Add check for autoconf paths.h value, include or define _PATH_TTY.
    Adding cast from char to int for all isspace/isdigit args.

bwctl.c:
    Adding cast from char to int for all isspace/isdigit args.
    Change setenv to putenv to set POSIXLY_CORRECT for getopt.
    Check for LOG_PERROR. (Add warning if -r is passed in, and
        LOG_PERROR is not available.)

bwctld.c:
    Add check for LOG_PERROR.

policy.c:
    Remove fts.h - not even used here, this is an owamp leftover...
    Adding cast from char to int for all isspace/isdigit args.

util.c:
    Adding cast from char to int for all isspace/isdigit args.


git-svn-id: https://svn.internet2.edu/svn/bwctl/trunk@294 2da7661c-602c-0410-9277-d9994c0c2ef1

+++ bwctld/policy.c
@@ -29,7 +29,6 @@
 #include <unistd.h>
 #include <netinet/in.h>
 #include <assert.h>
-#include <fts.h>
 
 #include <bwlib/bwlib.h>
 #include "policy.h"
@@ -254,9 +253,9 @@ parselimitline(
 		/*
 		 * Grab the keyname off the front.
 		 */
-		while(isspace(*line)){line++;}
+		while(isspace((int)*line)){line++;}
 		limname = line;
-		while(!isspace(*line) && (*line != '=')){
+		while(!isspace((int)*line) && (*line != '=')){
 			line++;
 		}
 		*line++ = '\0';
@@ -264,11 +263,11 @@ parselimitline(
 		/*
 		 * Grab the valname
 		 */
-		while(isspace(*line) || (*line == '=')){
+		while(isspace((int)*line) || (*line == '=')){
 			line++;
 		}
 		limval = line;
-		while(!isspace(*line) && (*line != '\0')){
+		while(!isspace((int)*line) && (*line != '\0')){
 			line++;
 		}
 		*line = '\0';
@@ -714,7 +713,7 @@ parselimits(
 		 */
 		if(!strncasecmp(line,"limit",5)){
 			line += 5;
-			while(isspace(*line)){
+			while(isspace((int)*line)){
 				line++;
 			}
 
@@ -729,7 +728,7 @@ parselimits(
 		 */
 		else if(!strncasecmp(line,"assign",6)){
 			line += 6;
-			while(isspace(*line)){
+			while(isspace((int)*line)){
 				line++;
 			}
